requirements.txt
python-telegram-bot==20.7
openai==1.30.1
python-dotenv
requests
PyMuPDF  # вместо fitz! Это официальный пакет для работы с PDF
pytesseract
Pillow
fastapi
uvicorn
watchdog
pdf2image==1.17.0  # Для конвертации PDF в картинки, нужен poppler-utils (ставится отдельно)
# openai-agents==0.1.6   # Если не используешь — удали или закомментируй
# httpx==0.27.0          # Обычно ставится как зависимость, не всегда нужен отдельно